refactor(network/diff): slightly reorganize diff
